Move news cell config method to extension.
